Add default parameter to `fn serialize`
This allows us to skip serializing values if they are equal to their
default value. The parameter is not supported for all Elements (this is
by design!), so assertions should be used liberally wherever they're not
supported (e.g. EBML Master Elements, UUIDs).

If the serialize function is used by API consumers, it should generally
be called with default set to None. Maybe there's a way to do this more
cleanly, without exposing the default parameter to API consumers?